requirements.txt
Flask
Flask-Cors
requests
Pillow
python-dotenv
openai